requirements.txt
flake8
flake8-isort
isort
black
autoflake
pytest
Flask
celery
flower
redis
google-cloud-storage
